fix(bedrock): update Mob::knockback and Player::startSleepInBed for 1.26.40
knockback's damage parameter changed from int to float. Arity and order
are otherwise identical - the old int was converted with cvtsi2ss inside
the callee, and 1.26.40 has no cvtsi2ss at all while a caller now passes
a float zero via xorps rather than an integer zero.

startSleepInBed gained a bool and a float after the block position. The
bool gates the respawn-point update, and the float replaces a constant
the previous build hardcoded when computing the sleep position. Their
names are guesses, so they are a2/a3 rather than invented descriptive
names; the types are certain.

Both mangled names change, so the config keys move with the
declarations. On Linux only knockback's key changes, since Itanium
omits return types. Both patterns already matched and are untouched.
Both platforms still resolve 61/61.

+++ src/bedrock/world/actor/mob.h
@@ -43,7 +43,7 @@ public:
     Mob(Level &, EntityContext &);
 
     ~Mob() override = 0;
-    ENDSTONE_HOOK virtual void knockback(Actor *, int damage, float xd, float zd,
+    ENDSTONE_HOOK virtual void knockback(Actor *, float damage, float xd, float zd,
                                          const KnockbackParameters &parameters);
     virtual void spawnAnim() = 0;
     virtual void setSprinting(bool) = 0;
